--- src/components/BoardDetail.tsx
@@ -1,5 +1,0 @@
-const BoardDetail = () => {
-  return;
-};
-
-export default BoardDetail;
