+++ src/components/home/ModalBoard.tsx
@@ -1,6 +1,8 @@
 import { useEffect, useState } from 'react';
 import ReactDOM from 'react-dom';
-import { ModalContainer } from '../common/styles';
+
+// styles
+import { ModalContainer } from '../../styles/common/styles';
 
 // icons
 import { IoClose } from 'react-icons/io5';
