+++ src/hooks/home/api/useDeleteBoard.ts
@@ -1,5 +1,10 @@
+// apis
 import { deleteBoardData } from '@/pages/api/clients/home';
+
+// types
 import { BoardType } from '@/types/home';
+
+// libraries
 import {
   QueryObserverResult,
   RefetchOptions,
