+++ src/hooks/home/api/useGetMyIndividualChat.ts
@@ -1,0 +1,23 @@
+// apis
+import { getMyChat } from '@/pages/api/clients/home';
+
+// types
+import { ChatDataType } from '@/types/chat';
+
+// libraries
+import { useQuery } from '@tanstack/react-query';
+import Cookies from 'js-cookie';
+
+const useGetMyIndividualChat = () => {
+  return useQuery<ChatDataType[]>({
+    queryKey: ['getMyIndividualChat'],
+    queryFn: async () => {
+      const chat_user_id = Number(Cookies.get('user_index'));
+      const response = await getMyChat(chat_user_id);
+
+      return response.data;
+    },
+  });
+};
+
+export default useGetMyIndividualChat;
