+++ src/hooks/home/api/useGetSpecificBoardData.ts
@@ -1,5 +1,10 @@
+// apis
 import { getSpecificBoard } from '@/pages/api/clients/home';
+
+// types
 import { BoardType } from '@/types/home';
+
+// libraries
 import { useQuery } from '@tanstack/react-query';
 
 interface useGetSpecificBoardDataType {
