+++ src/hooks/home/api/usePostBoardWrite.ts
@@ -1,5 +1,10 @@
+// apis
 import { postBoardData } from '@/pages/api/clients/home';
+
+// types
 import { BoardType } from '@/types/home';
+
+// libraries
 import {
   QueryObserverResult,
   RefetchOptions,
